refactor(utilities): introduce PathUtilities for organization name extraction and sanitization
